requirements.txt
py-postgresql
face_recognition
opencv-python
flask_restx
gunicorn